// Y-Recorder/src/models/observers.ts
export type Observer = {
  options?: boolean | object

  install(options?: object): void
  uninstall(): void
} & PubSubPattern

export type PubSubPattern = {
  queues: Map<string, Function[]>
  $on(hook: string, action: Function): void
  $off(hook: string, thisAction: Function): void
  $emit(hook: string, ...args): void
}

export interface Record {
  t: number // time
}

export enum ConsoleLevels {
  info = 'info',
  error = 'error',
  log = 'log',
  warn = 'warn',
  debug = 'debug'
}

export enum ConsoleTypes {
  console = 'console'
}

export type ConsoleRecord = {
  type: 'console'
  level: ConsoleLevels // console level,
  input?: any[]
}

export enum EventTypes {
  scroll = 'scroll',
  resize = 'resize',
  form = 'form'
}
// ...
export type Listener = {
  target: Window | Document | ElementX
  event: string
  callback: EventListenerOrEventListenerObject
  options?: AddEventListenerOptions | boolean
}

export type EventReocrd = {
  type: EventTypes
  // scroll
  x?: number
  y?: number
  // resize
  w?: number
  h?: number
  // form change
  target?: number
  tagName?:string
  k?: string
  v?: number | string
}